pyproject.toml
[tool.poetry]
name = "meal-advisor"
version = "0.1.0"
description = ""
authors = ["maxielias <maximilianocoltro@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
numpy = "^1.23.4"
pandas = "^1.5.1"
psycopg2-binary = "^2.9.5"
SQLAlchemy = "^1.4.44"
pytest = "^7.2.0"
streamlit = "^1.15.0"
aiohttp = "^3.8.3"
cchardet = "^2.1.7"
aiodns = "^3.0.0"
asyncio = "^3.4.3"
aiofiles = "^22.1.0"
ipykernel = "^6.17.1"
mealdb = "^1.0.0"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
lxml = "^4.9.1"
html5lib = "^1.1"
HTMLParser = "^0.0.2"
nutritionix = "^0.2"
streamlit-lottie = "^0.0.3"
Pillow = "^9.3.0"
plotly = "^5.11.0"
nltk = "^3.7"
ingredient-parser-nlp = "^0.1.0-alpha.3"
st-annotated-text = "^3.0.0"
streamlit-extras = "^0.2.4"
Google-Images-Search = "^1.4.6"
st-card-component = "^0.0.10"
PyAutoGUI = "^0.9.53"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"
ipykernel = "^6.17.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
